Some Changes

+++ controllers/product.js
@@ -56,9 +56,13 @@ const deleteProductById = async (req, res) => {
     try {
         let product = await Product.findByIdAndDelete(id);
         if (product) {
-            let oneStepBack = path.join(__dirname, '../');
-            await fs.unlinkSync(path.resolve(oneStepBack + 'public/images/' + product.imageURL));
+            if (!product.imageURL === "https://shorturl.at/RnRet") {    
+                let oneStepBack = path.join(__dirname, '../');
+                await fs.unlinkSync(path.resolve(oneStepBack + 'public/images/' + product.imageURL));
+            }
+            else{
             res.status(200).json({ success: true, message: "Product Deleted Successfully!" });
+            }
         } else {
             res.status(404).json({ success: false, message: "Product Not Found!" });
         }
